+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **BREAKING CHANGES**
 This version brings breaking changes.
+A [migration guide](https://adrien-bon.github.io/bevy_ecs_tiled/migrations/v0_12.html) is available to help adapt to these.
 
 ### Features